Adds a bit of entropy early in boot.
Right now, just the cycle time and timestamp for boot. These are
pretty attacker-guessable, but better than nothing.

This will be used to generate XOR cookies to initialize the heap with.

We usually get better entropy from U-Boot once we process the
Devicetree, so the RNG will reseed at that point.

Signed-off-by: Nathan Ringo <me@remexre.com>