Adds a revamped Devicetree parser.
This heap-allocates the Devicetree, so we don't need to keep the FDT
structure reserved.

It also includes the memory reservations block in the Devicetree proper,
under the /reserved-memory node.

This still needs to be fully integrated; e.g. by:

- Adding RAM to the allocators.
- Adding /chosen/rng-seed to the entropy pool.
- Creating devices from nodes.

This has been tested and works on the Milk-V Duo S.

Signed-off-by: Amy Ringo <me@remexre.com>